Change Gem name to rack-dev-insight

+++ gem/tasks/dummy_app/helper/setup_rails_helper.rb
@@ -2,20 +2,28 @@
 
 module SetupRailsHelper
   def stop_docker
+    puts '== Stop docker =='
     system('docker compose stop dummy-app-rails')
   end
 
   def build_image(version)
-    version ||= '3.1.3'
+    puts '== Build image =='
+    version ||= '3.0.2'
     system("RUBY_VERSION=#{version} docker compose build dummy-app-rails")
   end
 
-  def create_rails_app(version)
+  def create_rails_app_if_not_exists(version)
     version ||= '7.0'
+    if Dir.exist?('tmp/dummy_app/rails') && !Dir.empty?('tmp/dummy_app/rails')
+      puts '== Rails application already exists. Skip creating rails app. =='
+      return
+    end
+
+    puts "== Create rails app with version #{version} =="
     system(<<~COMMAND)
       docker compose run -it --rm dummy-app-rails /bin/bash -c '\
       gem install rails -v #{version} && \
-      rails new . --minimal;'
+      rails new . --minimal --skip-bundle;'
     COMMAND
   end
 
@@ -25,24 +33,39 @@ module SetupRailsHelper
       raise ArgumentError, "database must be one of 'sqlite', 'mysql', or 'pg'"
     end
 
+    puts "== Set database config to #{database} =="
     system("cp -f tasks/dummy_app/template_files/database-#{database}.yml tmp/dummy_app/rails/config/database.yml")
   end
 
-  def up_docker
-    system('docker compose up -d dummy-app-rails')
-  end
-
   def add_gems
+    puts '== Add gems =='
     system(<<~COMMAND)
-      docker compose exec dummy-app-rails /bin/bash -c '\
+      docker compose run -it --rm dummy-app-rails /bin/bash -c '\
+      bundle install && \
       (bundle show mysql2 || bundle add mysql2); \
       (bundle show pg || bundle add pg); \
       (bundle show net-http || bundle add net-http); \
-      (bundle show rack-analyzer || bundle add rack-analyzer --path /gem);'
+      (bundle show rack-dev-insight || bundle add rack-dev-insight --path /gem);'
     COMMAND
   end
 
+  def compile_gem
+    puts '== Compile gem =='
+    system(<<~COMMAND)
+      docker compose run -it --rm dummy-app-rails /bin/bash -c '\
+      cd /gem && \
+      bundle install && \
+      bundle exec rake compile;'
+    COMMAND
+  end
+
+  def up_docker
+    puts '== Up docker =='
+    system('docker compose up -d dummy-app-rails')
+  end
+
   def generate_scaffold
+    puts '== Generate scaffold =='
     system(<<~COMMAND)
       docker compose exec dummy-app-rails /bin/bash -c '\
       bundle install && \
@@ -51,6 +74,7 @@ module SetupRailsHelper
   end
 
   def migrate_reset
+    puts '== Migrate reset =='
     system(<<~COMMAND)
       docker compose exec dummy-app-rails /bin/bash -c '\
       bundle exec rails db:migrate:reset;'
@@ -58,6 +82,7 @@ module SetupRailsHelper
   end
 
   def restart_docker
+    puts '== Restart docker =='
     system('docker compose restart dummy-app-rails')
   end
 end
